Moving some code to separate segments.py module, adding searchf-test entry point
