+++ searchf/__init__.py
@@ -1,0 +1,1 @@
+__VERSION__ = '1.1'
