+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = searchf
-version = 1.0
+version = attr: searchf.__VERSION__
 description = A text-based application running in terminals to search into text files
 long_description = file: README.md
 long_description_content_type =	text/markdown
@@ -9,10 +9,10 @@ author = Adrien Felon
 url = https://github.com/human3/searchf
 license = Apache 2.0
 [options]
-package_dir =
-    = src
 packages = find:
-scripts =
-    bin/searchf
-[options.packages.find]
-where = src
+[options.package_data]
+* = *.md, *.txt
+[options.entry_points]
+console_scripts:
+    searchf=searchf.app:main
+    searchf-test=searchf.test.all:main
